[Refactor] Moved utility classes into separate packages
