+++ src/main/java/controllers/ProjectInputController.java
@@ -9,10 +9,10 @@ import models.reminder.Reminder;
 import models.task.ITask;
 import models.task.Task;
 import repositories.ProjectRepository;
-import util.AssignmentViewHelper;
-import util.CommandHelper;
+import util.uiformatter.AssignmentViewHelper;
+import util.uiformatter.CommandHelper;
 import util.ParserHelper;
-import util.ViewHelper;
+import util.uiformatter.ViewHelper;
 import util.date.DateTimeHelper;
 import util.factories.MemberFactory;
 import util.factories.ReminderFactory;
@@ -266,7 +266,7 @@ public class ProjectInputController implements IController {
      * Can be updated later on to include more information (tasks etc).
      * @param projectToManage The project specified by the user.
      */
-    public String[] projectViewMembers(Project projectToManage) {
+    private String[] projectViewMembers(Project projectToManage) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[projectViewMembers]");
         ArrayList<String> allMemberDetailsForTable = projectToManage.getMembers().getAllMemberDetailsForTable();
         String header = "Members of " + projectToManage.getName() + ":";
@@ -281,7 +281,7 @@ public class ProjectInputController implements IController {
      * Displays the members’ credits, their index number, name, and name of tasks completed.
      * @param projectToManage The project specified by the user.
      */
-    public String[] projectViewCredits(Project projectToManage) {
+    private String[] projectViewCredits(Project projectToManage) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[projectViewCredits]");
         ArrayList<String> allCredits = projectToManage.getCredits();
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "allCredits: " + allCredits.toString());
@@ -421,7 +421,7 @@ public class ProjectInputController implements IController {
      * @param projectToManage The project specified by the user.
      * @param projectCommand The user input.
      */
-    public String[] projectViewTaskRequirements(Project projectToManage, String projectCommand) {
+    private String[] projectViewTaskRequirements(Project projectToManage, String projectCommand) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(),
                 "[projectViewTaskRequirements] User input: '" + projectCommand + "'");
         if (projectCommand.length() < 23) {
@@ -452,7 +452,7 @@ public class ProjectInputController implements IController {
      * @param projectToManage The project specified by the user.
      * @param projectCommand The user input.
      */
-    public String[] projectViewTasks(Project projectToManage, String projectCommand) {
+    private String[] projectViewTasks(Project projectToManage, String projectCommand) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[projectViewTasks] User input: '"
             + projectCommand + "'");
         try {
@@ -543,7 +543,7 @@ public class ProjectInputController implements IController {
      * @param projectToManage the project being managed.
      * @param projectCommand The command by the user containing index numbers of the members to view.
      */
-    public String[] projectViewMembersAssignments(Project projectToManage, String projectCommand) {
+    private String[] projectViewMembersAssignments(Project projectToManage, String projectCommand) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(),
                 "[projectViewMembersAssignments] User input: '" + projectCommand + "'");
         ParserHelper parserHelper = new ParserHelper();
@@ -580,7 +580,7 @@ public class ProjectInputController implements IController {
      * @param projectToManage The project specified by the user.
      * @return Boolean variable specifying the exit status.
      */
-    public String[] projectExit(Project projectToManage) {
+    private String[] projectExit(Project projectToManage) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[projectExit]");
         return new String[] {"Exited project: " + projectToManage.getName()};
     }
@@ -589,7 +589,7 @@ public class ProjectInputController implements IController {
         return isManagingAProject;
     }
 
-    public String[] end() {
+    private String[] end() {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[end]");
         return new String[] {"Bye. Hope to see you again soon!"};
     }
@@ -624,7 +624,7 @@ public class ProjectInputController implements IController {
      * View reminder to the default list list of tasks and the members assigned to them.
      * @param projectToManage The project to manage.
      */
-    public String [] projectViewReminder(Project projectToManage) {
+    private String [] projectViewReminder(Project projectToManage) {
         ArchDukeLogger.logDebug(ProjectInputController.class.getName(), "[projectViewReminder]");
         DateTimeHelper dateTimeHelper = new DateTimeHelper();
         ArrayList<ArrayList<String>> tableToPrint = new ArrayList<>();
