+++ src/main/java/models/member/MemberList.java
@@ -1,7 +1,7 @@
 package models.member;
 
 import util.ParserHelper;
-import util.ValidityHelper;
+import util.validation.ValidityHelper;
 
 import java.util.ArrayList;
